{"types":"C","runs":17}

+++ MidTerm191.java
@@ -2,7 +2,7 @@ import acm.program.ConsoleProgram;
 
 public class MidTerm191 extends ConsoleProgram{
 	public void run(){
-		double a =(double) 12 / 5 + 2.4/2 + 2/5;
+		double a =(double) 12 / 5 + (double)2.4/2 + (double)2/5;
 		println(a);
 	}
 }
